Graduate the integrated-transport-solver mixin files to strict cohort
Complete the transport-solver cluster: extend TransportSolverState with the remaining
cross-mixin surface (n_D/n_T/n_He profiles, T_edge_keV, neoclassical_params, the neural-model
and OOD-selection helpers, BoolArray), type the previously untyped runtime kernels
(_thomas_solve, _explicit_diffusion_rhs, _build_cn_tridiag) which clears the downstream
has-type chain, give the eight mixin files float64 array typing (np.ndarray -> FloatArray,
no-any-return coercions on the external solver-module returns, the OOD mask bool return, the
nan_to_num overload) and add them to the cohort. With this the strict mypy + numpy-docstring
cohort covers every src module — non-cohort is zero. Full strict gate Success on 297 files;
116 transport tests pass (no MRO runtime shadow). One numpy-1.26.4-only float coercion
(np.min) was caught by the faithful CI mirror.

Authored by Anulum Fortis & Arcane Sapience (protoscience@anulum.li)

+++ src/scpn_fusion/core/_integrated_transport_solver_init.py
@@ -23,7 +23,7 @@ class TransportSolverInitializationMixin(TransportSolverState):
 
         self.external_profile_mode = True
         self.nr = 50
-        self.rho = np.linspace(0, 1, self.nr)
+        self.rho = np.linspace(0, 1, self.nr, dtype=np.float64)
         self.drho = 1.0 / (self.nr - 1)
 
         self.multi_ion: bool = multi_ion
